Update streamlit_app.py
Connecting your app to Snowflake is less easy in SniS than it was in SiS
With SniS, users will be able to connect to your app more easily. You can set up your SniS app in a way that doesn't require them to log in or have a USER in your Snowflake account. In fact, Streamlit will host your app for free if you make it open to the public.